requirements.txt
psycopg2-binary>=2.9
python-dotenv>=1.0
fastapi>=0.111
uvicorn[standard]>=0.29
